requirements.txt
python-telegram-bot[asyncio]==22.2
SQLAlchemy==2.0.41
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiohttp==3.12.14
aiocache==0.12.3
cachetools==6.1.0
python-dotenv==1.1.1
watchfiles==1.1.0
alembic==1.16.4
humanize==4.12.3
